Fix: separate models and interfaces
